+++ relay-server/src/utils/dynamic_sampling.rs
@@ -101,6 +101,41 @@ pub fn get_sampling_result(
     SamplingResult::determine_from_sampling_match(sampling_result)
 }
 
+/// Runs dynamic sampling if the dsc and root project state are not None and returns whether the
+/// transactions received with such dsc and project state would be kept or dropped by dynamic
+/// sampling.
+pub fn is_trace_fully_sampled(
+    processing_enabled: bool,
+    root_project_state: Option<&ProjectState>,
+    dsc: Option<&DynamicSamplingContext>,
+) -> Option<bool> {
+    let dsc = dsc?;
+    let root_project_state = root_project_state?;
+
+    // If the sampled field is not set, we prefer to not tag the error since we have no clue on
+    // whether the head of the trace was kept or dropped on the client side.
+    // In addition, if the head of the trace was dropped on the client we will immediately mark
+    // the trace as not fully sampled.
+    if !(dsc.sampled?) {
+        return Some(false);
+    }
+
+    let sampling_result = get_sampling_result(
+        processing_enabled,
+        None,
+        Some(root_project_state),
+        Some(dsc),
+        None,
+    );
+
+    let sampled = match sampling_result {
+        SamplingResult::Keep => true,
+        SamplingResult::Drop(_) => false,
+    };
+
+    Some(sampled)
+}
+
 /// Returns the project key defined in the `trace` header of the envelope.
 ///
 /// This function returns `None` if:
@@ -152,6 +187,7 @@ mod tests {
         release: Option<&str>,
         transaction: Option<&str>,
         environment: Option<&str>,
+        sampled: Option<bool>,
     ) -> DynamicSamplingContext {
         DynamicSamplingContext {
             trace_id: Uuid::new_v4(),
@@ -163,6 +199,7 @@ mod tests {
             user: Default::default(),
             other: Default::default(),
             replay_id: None,
+            sampled,
         }
     }
 
@@ -260,9 +297,68 @@ mod tests {
             rules_v2: vec![mocked_sampling_rule(1, RuleType::Trace, 1.0)],
             mode: SamplingMode::Received,
         });
-        let dsc = mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None);
+        let dsc = mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None, None);
 
         let result = get_sampling_result(true, None, Some(&root_project_state), Some(&dsc), None);
         assert_eq!(result, SamplingResult::Keep)
+    }
+
+    #[test]
+    /// Tests that a trace is marked as fully sampled correctly when dsc and project state are set.
+    fn test_is_trace_fully_sampled_with_valid_dsc_and_project_state() {
+        // We test with `sampled = true` and 100% rule.
+        let project_state = project_state_with_config(SamplingConfig {
+            rules: vec![],
+            rules_v2: vec![mocked_sampling_rule(1, RuleType::Trace, 1.0)],
+            mode: SamplingMode::Received,
+        });
+        let dsc =
+            mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None, Some(true));
+
+        let result = is_trace_fully_sampled(true, Some(&project_state), Some(&dsc)).unwrap();
+        assert!(result);
+
+        // We test with `sampled = true` and 0% rule.
+        let project_state = project_state_with_config(SamplingConfig {
+            rules: vec![],
+            rules_v2: vec![mocked_sampling_rule(1, RuleType::Trace, 0.0)],
+            mode: SamplingMode::Received,
+        });
+        let dsc =
+            mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None, Some(true));
+
+        let result = is_trace_fully_sampled(true, Some(&project_state), Some(&dsc)).unwrap();
+        assert!(!result);
+
+        // We test with `sampled = false` and 100% rule.
+        let project_state = project_state_with_config(SamplingConfig {
+            rules: vec![],
+            rules_v2: vec![mocked_sampling_rule(1, RuleType::Trace, 1.0)],
+            mode: SamplingMode::Received,
+        });
+        let dsc =
+            mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None, Some(false));
+
+        let result = is_trace_fully_sampled(true, Some(&project_state), Some(&dsc)).unwrap();
+        assert!(!result);
+    }
+
+    #[test]
+    /// Tests that a trace is not marked as fully sampled or not if inputs are invalid.
+    fn test_is_trace_fully_sampled_with_invalid_inputs() {
+        // We test with missing `sampled`.
+        let project_state = project_state_with_config(SamplingConfig {
+            rules: vec![],
+            rules_v2: vec![mocked_sampling_rule(1, RuleType::Trace, 1.0)],
+            mode: SamplingMode::Received,
+        });
+        let dsc = mocked_simple_dynamic_sampling_context(Some(1.0), Some("3.0"), None, None, None);
+
+        let result = is_trace_fully_sampled(true, Some(&project_state), Some(&dsc));
+        matches!(result, None);
+
+        // We test with missing dsc and project config.
+        let result = is_trace_fully_sampled(true, None, None);
+        matches!(result, None);
     }
 }